Private Constructor & Comment
Prevent instantiation of static class.
Remove trailing comment that served no purpose.  Please note that if the
intent was to not print the remainder of the next line, it was not
actually doing that and was printing it.  So this just removes the
ending line comment.  It does appear the entire thing needed printed.