[ci] Run whitespace-maven-plugin to trim excessive space
no-op action.  This trims all trailing space which results in little to see on git.  This is another attempt at aligning code for future locking of formatting into builds.  This plugin can do more than simply java files so it has cleared issues in xml and some deeper issues in at least two of the site pages which seemed to cause license header to post twice.