Report mapper methods with no SQL statement to bind
Currently, an unbound method (i.e. mapper method that does not have SQL statement to bind) causes an error only when the method is invoked.
With this change, the error is reported when invoking a statement (any statement) for the first time.
There is no false-positives that I am aware.