Pooled connection timeout in pingConnection method
If the firewall drops packets or they are lost on the network, the pingConnection method in X ms drops this connection and gets a new one, and does not wait indefinitely the response. The time will be configured in networkPingTimeout property (in milliseconds).
To reproduce this execute this in the database machine:
 sudo iptables -D INPUT -p tcp --sport $i -j DROP; sudo iptables -D OUTPUT -p tcp --dport $i -j DROP
 $i -> port of the connections