Type handler should be applied to args for nested select
For "composite keys", the behavior changes depending on the existence of `parameterType`
When `parameterType` is not specified, parameter values are stored in a `Map` and custom type handler is not used.
When `parameterType="User"` is specified, `User` instance is created and properties are set using type handlers.
Related to gh-114

For "simple key", `parameterType` is ignored and custom type handler is not used at all.

I can think of some corner cases caused by these behaviors, but will deal with them when/if they come up.